save cffs in memory using bitmap

+++ src/cff_file_generator.h
@@ -1,6 +1,8 @@
 #ifndef CFF_FILE_GENERATOR_H
 #define CFF_FILE_GENERATOR_H
 
+#include <stdint.h>
+
 struct cff_parameters {
     char construction;
     int* Fqs;
@@ -9,8 +11,8 @@ struct cff_parameters {
     int ks_count; 
 };
 
-int** read_cff_from_file(const char* filename, long* rows, long* cols);
-void write_cff_to_file(const char* filename, char construction, long* Fq_steps, int fqs_count, long* K_steps, int ks_count,  int** matrix, long rows, long cols);
+uint64_t** read_cff_from_file(const char* filename, long* rows, long* cols);
+void write_cff_to_file(const char* filename, char construction, long* Fq_steps, int fqs_count, long* K_steps, int ks_count,  uint64_t** matrix, long rows, long cols);
 struct cff_parameters* read_parameters(const char* filename);
 
 #endif
